Add files via upload
Writing to digital pins, reading from digital pins, and how to use i2c RTC with an i2c LED Screen.
RTC     
An Arduino's memory is wiped when it loses power. For most of our projects, that doesn't matter b/c we're only using the Arduino as a momentary I/O controller. However, when we have a project that needs to be persistent, we need a way to keep time even when the Arduino is powered off.        

Segment Summary:        
    We're going to connect a Tiny RTC to an Arduino.        
Parts:        
    1. Tiny RTC I2C DS1307 Real Time Clock  $2.00    
    2. Arduino        
    3. Jumper Wires        
What is I2C?        
    * I2C (Pronounced "I Squared C") can communicate with multiple devices over short distances with just two wires: SDA (Data Line) & SCL (Clock Line)        
    * Technically, it's a "Multi-Master, Multi-Slave, Single-Ended, Serial bus"        
            
You need to know what are the SCL and SDA pins on your arduino        
    * Uno, Nano or clone